# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CampaignCreateRequest": {
        "description": "建立活動（DRAFT）。寄售折扣預設關；品項預設自有序號+自有散裝開（docs/21 §8）。",
        "properties": {
          "applies_catalog": {
            "default": false,
            "title": "Applies Catalog",
            "type": "boolean"
          },
          "applies_consignment": {
            "default": false,
            "title": "Applies Consignment",
            "type": "boolean"
          },
          "applies_owned_bulk": {
            "default": true,
            "title": "Applies Owned Bulk",
            "type": "boolean"
          },
          "applies_owned_serialized": {
            "default": true,
            "title": "Applies Owned Serialized",
            "type": "boolean"
          },
          "consignment_discount_bearing": {
            "$ref": "#/components/schemas/ConsignmentDiscountBearing",
            "default": "STORE_ABSORBS"
          },
          "discount_pct": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Discount Pct",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "discount_pct",
          "starts_at",
          "ends_at"
        ],
        "title": "CampaignCreateRequest",
        "type": "object"
      },
      "CampaignRead": {
        "properties": {
          "applies_catalog": {
            "title": "Applies Catalog",
            "type": "boolean"
          },
          "applies_consignment": {
            "title": "Applies Consignment",
            "type": "boolean"
          },
          "applies_owned_bulk": {
            "title": "Applies Owned Bulk",
            "type": "boolean"
          },
          "applies_owned_serialized": {
            "title": "Applies Owned Serialized",
            "type": "boolean"
          },
          "consignment_discount_bearing": {
            "$ref": "#/components/schemas/ConsignmentDiscountBearing"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "integer"
          },
          "discount_pct": {
            "title": "Discount Pct",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          },
          "store_id": {
            "title": "Store Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "store_id",
          "name",
          "discount_pct",
          "applies_owned_serialized",
          "applies_owned_bulk",
          "applies_catalog",
          "applies_consignment",
          "consignment_discount_bearing",
          "starts_at",
          "ends_at",
          "status",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignRead",
        "type": "object"
      },
      "CampaignStatus": {
        "description": "門市活動狀態機（docs/21）。DRAFT→ACTIVE→ENDED；DRAFT/ACTIVE→CANCELLED。",
        "enum": [
          "DRAFT",
          "ACTIVE",
          "ENDED",
          "CANCELLED"
        ],
        "title": "CampaignStatus",
        "type": "string"
      },
...
      "ConsignmentDiscountBearing": {
        "description": "活動折扣套用寄售品時，折讓由誰吸收（docs/21 §8.1；applies_consignment=true 時生效）。\n\nSTORE_ABSORBS：寄售人 payout 認原 listed_price、抽成吸收折讓（且抽成不得 < 0）。\nPROPORTIONAL：gross=折後價，抽成與 payout 按折後縮水（等於替寄售人打折）。",
        "enum": [
          "STORE_ABSORBS",
          "PROPORTIONAL"
        ],
        "title": "ConsignmentDiscountBearing",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "lu-camp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "operationId": "listCampaigns",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CampaignStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignRead"
                  },
                  "title": "Response Listcampaigns",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Campaigns",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "createCampaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}": {
      "get": {
        "operationId": "getCampaign",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/activate": {
      "post": {
        "operationId": "activateCampaign",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Activate Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/cancel": {
      "post": {
        "operationId": "cancelCampaign",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/end": {
      "post": {
        "operationId": "endCampaign",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "End Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  }
}